examples/service/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6
django-cors-headers==3.7.0
drf-yasg==1.20.0
graphene-django==2.15.0
djangorestframework-simplejwt==4.6.0
